+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Implemented Primary IPs
 - Allow calling `#update` with additional attributes to be updated
 - Return the resource when calling `#create`, `#update` or `#delete`
 - Add additional primary IP mismatch errors: